requirements.txt
Flask
pymongo
requests
gunicorn
google-genai
google-generativeai
python-dotenv